Extact "impl Debug for *Entry" code into common/map.rs
