+++ src/weak_value_hash_map.rs
@@ -494,35 +494,11 @@ where
     }
 }
 
-impl<'a, K: Debug, V: WeakElement> Debug for Entry<'a, K, V>
-where
-    V::Strong: Debug,
-{
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        match *self {
-            Entry::Occupied(ref e) => e.fmt(f),
-            Entry::Vacant(ref e) => e.fmt(f),
-        }
-    }
-}
-
-impl<'a, K: Debug, V: WeakElement> Debug for OccupiedEntry<'a, K, V>
-where
-    V::Strong: Debug,
-{
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl<'a, K: Debug, V: WeakElement> Debug for VacantEntry<'a, K, V>
-where
-    V::Strong: Debug,
-{
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
+debug_for_entry! {where {
+    K: Debug,
+    V: WeakElement,
+    V::Strong: Debug
+}}
 
 impl<K, V: WeakElement, S> IntoIterator for WeakValueHashMap<K, V, S> {
     type Item = (K, V::Strong);
